+++ pkg/services/import_test.go
@@ -11,7 +11,7 @@ func Test_convertToMilliseconds(t *testing.T) {
 	tests := []struct {
 		name string
 		args args
-		want int
+		want int64
 	}{
 		{
 			name: "test",
